Reorganize and Rename to have multiple autonomy algs (#107)
* precommit

* add non-descriptive docstrings :)

* reorganize control

* add non-descriptive docstrings

* snakeCase pid controller

* SnakeCase files pid

* trying to add utils (and failing)

* Revert "trying to add utils (and failing)", decided to save utils for a future commit

This reverts commit c2883d155584bd60e3ed7857d7a8b66b660695ed.

* seperate ekf

* add shared utils for localization

* fix publisher, remove unneeded import

* add particle filter, unsure how to deal iwth launchfiles...

* finish SE

* rename path planner

* add seperate launch descriptions, I don't think I did this the best way though (works for now though....)

* rename pid_controller to pid_lateral_controller

* rename particle filter

* rename localization_shared_utils

* fix Aaron suggestions

* add disable launch argument
