+++ workspace/src/common/launch/art_control_launch/package.xml
@@ -9,7 +9,7 @@
 
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
 
-  <exec_depend>control</exec_depend>
+  <exec_depend>pid_lateral_controller</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
