--- workspace/src/control/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>control</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="amelmquist@wisc.edu">asher</maintainer>
-  <license>TODO: License declaration</license>
-
-  <exec_depend>rclpy</exec_depend>
-  <exec_depend>ament_index_python</exec_depend>
-
-  <exec_depend>art_msgs</exec_depend>
-  <exec_depend>nav_msgs</exec_depend>
-
-  <exec_depend>python3-numpy</exec_depend>
-  <exec_depend>tf_transformations</exec_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
