+++ workspace/src/control/pid_lateral_controller/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/pid_lateral_controller
+[install]
+install_scripts=$base/lib/pid_lateral_controller
