--- workspace/src/control/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/control
-[install]
-install_scripts=$base/lib/control
