+++ workspace/src/localization/ekf_estimation/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/ekf_estimation
+[install]
+install_scripts=$base/lib/ekf_estimation
