+++ workspace/src/localization/ground_truth/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/ground_truth
+[install]
+install_scripts=$base/lib/ground_truth
