+++ workspace/src/localization/ground_truth/setup.py
@@ -1,0 +1,23 @@
+from setuptools import setup
+
+package_name = "ground_truth"
+
+setup(
+    name=package_name,
+    version="0.0.0",
+    packages=[package_name],
+    data_files=[
+        ("share/ament_index/resource_index/packages", ["resource/" + package_name]),
+        ("share/" + package_name, ["package.xml"]),
+    ],
+    install_requires=["setuptools"],
+    zip_safe=True,
+    maintainer="art",
+    maintainer_email="art@todo.todo",
+    description="TODO: Package description",
+    license="TODO: License declaration",
+    tests_require=["pytest"],
+    entry_points={
+        "console_scripts": ["ground_truth = ground_truth.ground_truth:main"],
+    },
+)
