--- workspace/src/localization/localization_py/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/localization_py
-[install]
-install_scripts=$base/lib/localization_py
