+++ workspace/src/localization/localization_shared_utils/localization_shared_utils/__init__.py
@@ -1,0 +1,11 @@
+"""Main entry point for the `shared_utils` package"""
+
+from .localization_shared_utils import (
+    get_dynamics,
+    get_coordinate_transfer,
+)
+
+__all__ = [
+    "get_dynamics",
+    "get_coordinate_transfer",
+]
