+++ workspace/src/localization/particle_filter_estimation/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/particle_filter_estimation
+[install]
+install_scripts=$base/lib/particle_filter_estimation
