+++ workspace/src/path_planning/centerline_objects_path_planner/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/centerline_objects_path_planner
+[install]
+install_scripts=$base/lib/centerline_objects_path_planner
