--- workspace/src/path_planning/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>path_planning</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="amelmquist@wisc.edu">asher</maintainer>
-  <license>TODO: License declaration</license>
-
-  <exec_depend>rclpy</exec_depend>
-
-  <exec_depend>art_msgs</exec_depend>
-  <exec_depend>art_perception_msgs</exec_depend>
-  <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>nav_msgs</exec_depend>
-  <exec_depend>geometry_msgs</exec_depend>
-
-  <exec_depend>python3-numpy</exec_depend>
-  <exec_depend>python3-matplotlib</exec_depend>
-  <exec_depend>python3-scipy</exec_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
