--- workspace/src/path_planning/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/path_planning
-[install]
-install_scripts=$base/lib/path_planning
